add rng device; make sure new disk is larger than backing disk image
VMs may hang if there is not enough entropy for certain crypto operations.
Add a random device to the guest to provide entropy from the host rng.
This also requires that the guest install rng-tools and run rngd which
will make sure /dev/random in the guest is filled with entropy from
the virtual device from the host.
This is controlled by the config parameter VM_RNG - by default it is set
VM_RNG="--rng /dev/random"
When creating a new disk image based on a backing image, the size of the
new disk image must be greater than or equal to the size of the backing
image.  This version of the script ensures that will happen.