Correct backing file size comparison
The backing file size is compared with the requested size to
ensure that the requested size is larger than the backing file.
This comparison only handles integer values.  If there is a
fractional part in the backing file size, the comparison could be
performed incorrectly resulting in an image that is too small.

This patch drops off the fractional part of the backing file size
before performing the comparison.