Allow cloud-init to add yum repos and packages
The VM config files currently allow one to define yum repositories
and packages to install by using the following variables:

  - VM_OS_BASE_REPO_LIST
  - VM_REPO_LIST
  - VM_PACKAGE_LIST

Unfortuantely, these only work when using a kickstart.  This patch
allows these same variables to be used to update the cloud-config
data that is used by cloud-init.  This allows yum reposities, yum
updating, and installation of additional packages to be easily
defined in the VM config file instead of resorting to handling it
in the post-cloud-init shell script.