Allow additional network interfaces to be created
This patch allows secondary and tertiary network interfaces to be
created when running virt-install.  The VM config files can specify
the virtual network to attach the interface to by using the following
variables:

  VM_NETWORK_NAME_2
  VM_NETWORK_NAME_3

It is assumed that the virtual networks have already been created,
and no extra work is performed to add host info to the additional
virtual networks.